# api/openapi.yaml
openapi: 3.0.3
...
info:
  title: VibeShield API
  version: 1.0.0
  description: |
    Public REST contract for VibeShield — surface-level security scans for
    AI-built websites, with AI-generated fix prompts per finding.

    Clients reach this API same-origin through the nginx gateway (locally) or
    the Kubernetes ingress (deployed): `/api/v1/auth/*` is served by the
    auth-service, every other `/api/v1/*` path by the api-service. The split is
    a routing detail; consumers see one API.

    **Async scan workflow:** `POST /api/v1/websites/{websiteId}/scans` returns
    `202 Accepted` immediately with the scan in status `Pending`. Poll
    `GET /api/v1/scans/{scanId}` until the status is `Completed` or `Failed`,
    then fetch `GET /api/v1/scans/{scanId}/findings`.
...
servers:
  - url: /
    description: Same-origin via the nginx gateway / Kubernetes ingress.
security:
  - bearerAuth: []
paths:
...
  /api/v1/websites/{websiteId}/scans:
    parameters:
      - $ref: "#/components/parameters/WebsiteId"
    post:
      tags: [Scans]
      operationId: triggerScan
      summary: Trigger a new scan for a website
      description: |
        Creates the scan in status `Pending` and returns immediately; execution
        happens in the background. At most one scan per website may be
        `Pending` or `Running` at a time — a second trigger returns `409`.
      requestBody:
        required: false
        description: Optional scan configuration; omitting it runs all checks with defaults.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanRequest"
      responses:
        "202":
          description: Scan accepted for background execution.
          headers:
            Location:
              description: URL of the created scan, for status polling.
              schema:
                type: string
                example: /api/v1/scans/42
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: A scan for this website is already `Pending` or `Running` (`SCAN_IN_PROGRESS`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags: [Scans]
      operationId: listScans
      summary: List the scan history of a website
      responses:
        "200":
          description: All scans for the website, newest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scan"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/websites/{websiteId}/scans/latest:
    get:
      tags: [Scans]
      operationId: getLatestScan
      summary: Return the most recent scan of a website
      parameters:
        - $ref: "#/components/parameters/WebsiteId"
      responses:
        "200":
          description: The most recently created scan.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Website not visible to this user, or it has no scans yet (`SCAN_NOT_FOUND`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/scans/{scanId}:
    get:
      tags: [Scans]
      operationId: getScan
      summary: Return a scan's current status
      description: Poll this endpoint while the status is `Pending` or `Running`.
      parameters:
        - $ref: "#/components/parameters/ScanId"
      responses:
        "200":
          description: Current scan state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/scans/{scanId}/findings:
    get:
      tags: [Scans]
      operationId: listFindings
      summary: List the findings of a scan
      description: Empty until the scan status is `Completed`.
      parameters:
        - $ref: "#/components/parameters/ScanId"
      responses:
        "200":
          description: Findings ordered by severity, most severe first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Finding"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/scans/{scanId}/comparison:
    get:
      tags: [Scans]
      operationId: getScanComparison
      summary: Compare this scan with the previous completed scan
      description: Compares the selected completed scan with the latest previous completed scan for the same website.
      parameters:
        - $ref: "#/components/parameters/ScanId"
      responses:
        "200":
          description: Comparison result and prioritized action plan.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanComparison"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by `POST /api/v1/auth/login`.
  parameters:
    WebsiteId:
      name: websiteId
      in: path
      required: true
      description: ID of a website registered by the authenticated user.
      schema:
        type: integer
        format: int64
    ScanId:
      name: scanId
      in: path
      required: true
      description: ID of a scan belonging to one of the user's websites.
      schema:
        type: integer
        format: int64
  responses:
    Unauthorized:
      description: Missing, expired, or invalid Bearer token (`UNAUTHORIZED`, `INVALID_TOKEN`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource does not exist or belongs to another user (`WEBSITE_NOT_FOUND`, `SCAN_NOT_FOUND`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      description: Unified error shape returned by every VibeShield service.
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error code, e.g. `SCAN_IN_PROGRESS`.
          example: SCAN_IN_PROGRESS
        message:
          type: string
          description: Human-readable explanation.
          example: A scan for this website is already running.
        details:
          type: object
          nullable: true
          description: Optional structured context (e.g. field-level validation errors).
          additionalProperties: true
...
    ScanRequest:
      type: object
      description: Configuration for a scan run. All fields are optional.
      properties:
        checks:
          type: array
          description: Checks to run; defaults to all available checks.
          items:
            $ref: "#/components/schemas/ScanCheck"
        crawlDepth:
          type: integer
          minimum: 0
          maximum: 3
          description: How many link levels to follow from the start URL. Defaults to 0, which scans only the registered URL.
        includeSubdomains:
          type: boolean
          description: Whether the crawl may follow links to subdomains of the registered host. Defaults to false.
    Scan:
      type: object
      description: A scan job and its current state.
      required: [id, websiteId, status, createdAt]
      properties:
        id:
          type: integer
          format: int64
          example: 42
        websiteId:
          type: integer
          format: int64
          example: 7
        status:
          $ref: "#/components/schemas/ScanStatus"
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
          description: Set once the status is `Completed` or `Failed`.
        errorMessage:
          type: string
          nullable: true
          description: Why the scan failed; only set when the status is `Failed`.
        findingCount:
          type: integer
          nullable: true
          description: Number of findings; only set when the status is `Completed`.
          example: 5
    ScanStatus:
      type: string
      description: Lifecycle of a scan job.
      enum: [Pending, Running, Completed, Failed]
    ScanCheck:
      type: string
      description: |
        A surface-level check the scanner can run. Values match the scan
        options offered in the client UI.
      enum: [crawl, https, headers, adminPaths, secrets, sensitiveFiles]
    Severity:
      type: string
      description: How urgently a finding should be fixed.
      enum: [Critical, High, Medium, Low, Info]
    FindingStatus:
      type: string
      description: Triage state of a finding. New findings start as `Open`.
      enum: [Open, Fixed, Ignored]
    Finding:
      type: object
      description: A single security issue discovered by a scan.
      required: [id, scanId, check, title, severity, affected, explanation, suggestedFix, status]
      properties:
        id:
          type: integer
          format: int64
        scanId:
          type: integer
          format: int64
        check:
          $ref: "#/components/schemas/ScanCheck"
        title:
          type: string
          example: Missing Content-Security-Policy header
        severity:
          $ref: "#/components/schemas/Severity"
        affected:
          type: string
          description: URL, path, or header the finding applies to.
          example: https://my-vibecoded-shop.lovable.app/
        explanation:
          type: string
          description: Non-technical explanation of the risk, written for vibecoders.
        suggestedFix:
          type: string
          description: Plain-language fix suggestion; input for AI fix-prompt generation.
        status:
          $ref: "#/components/schemas/FindingStatus"
    ScanChangeStatus:
      type: string
      description: How a finding changed compared with the previous completed scan.
      enum: [Fixed, Still present, Newly introduced]
    ComparisonFinding:
      type: object
      required: [changeStatus, severity, check, title, affected, suggestedFix, effort]
      properties:
        findingId:
          type: integer
          format: int64
          nullable: true
          description: ID of the finding in the scan it came from.
        changeStatus:
          $ref: "#/components/schemas/ScanChangeStatus"
        severity:
          $ref: "#/components/schemas/Severity"
        check:
          $ref: "#/components/schemas/ScanCheck"
        title:
          type: string
        affected:
          type: string
        suggestedFix:
          type: string
        suggestedFixOrder:
          type: integer
          nullable: true
          description: 1-based action-plan order for current open findings.
        effort:
          $ref: "#/components/schemas/EffortEstimate"
    ScanComparisonSummary:
      type: object
      required: [fixed, stillPresent, newlyIntroduced]
      properties:
        fixed:
          type: integer
        stillPresent:
          type: integer
        newlyIntroduced:
          type: integer
    ScanComparison:
      type: object
      required: [scanId, comparable, message, summary, findings, actionPlan]
      properties:
        scanId:
          type: integer
          format: int64
        previousScanId:
          type: integer
          format: int64
          nullable: true
        comparable:
          type: boolean
          description: False when the scan is not completed or no previous completed scan exists.
        message:
          type: string
        summary:
          $ref: "#/components/schemas/ScanComparisonSummary"
        findings:
          type: array
          description: Fixed, still-present, and newly introduced findings.
          items:
            $ref: "#/components/schemas/ComparisonFinding"
        actionPlan:
          type: array
          description: Current open findings ordered by severity and rough effort estimate.
          items:
            $ref: "#/components/schemas/ComparisonFinding"
    EffortEstimate:
      type: object
      required: [level, estimate]
      properties:
        level:
          type: string
          enum: [Low, Medium, High]
        estimate:
          type: string
          example: 1-2 hours
...